mcp_usecases/16_github_claude/pyproject.toml
[project]
name = "16-github-claude"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.27.0",
    "python-dotenv>=1.2.2",
]
